Replace "[Nn]evermind" with "[Nn]ever mind"
Interestingly, some of the vanilla files (ENGLISH_BASE] also had this
issue. The other changes were mine though; even though nevermind is
wrong it looks better to me, especially because no one emphasizes a
pause when speaking. Everyone in real life says it as if it's spelled
"nevermind". Oh well. Thanks @Prisoner416 for pointing this out!