[EN_FXT] Merging last of robbforce changes
@robbforce Long story short, I had to delete the branch. I couldn't get
it to merge and was having all sorts of issues. I apologize if this
deletes the local files on your end. But I merged all the old stuff of
yours (MSG Review 8/9) from my local drive that I still hadn't merged
yet, so there's a lot of good changes in this and the last few commits,
almost all of them from you. :)